+++ src/main/java/wanted/media/post/dto/PostDto.java
@@ -1,11 +1,11 @@
-package wanted.media.content.dto;
+package wanted.media.post.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import wanted.media.content.domain.Post;
-import wanted.media.content.domain.Type;
+import wanted.media.post.domain.Post;
+import wanted.media.post.domain.Type;
 
 import java.time.LocalDateTime;
 
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class PostDto {
-    private Long id;
+    private String id;
     private Long likeCount;
     private Type type;
     private String title;
