+++ src/main/java/wanted/media/post/repository/PostRepository.java
@@ -5,12 +5,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import wanted.media.content.domain.Post;
-import wanted.media.content.domain.Type;
+import wanted.media.post.domain.Post;
+import wanted.media.post.domain.Type;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
     @Query("SELECT p FROM Post p " +
-            "WHERE p.user.account LIKE %:account% " +
+            "WHERE p.user.account = :account " +
             "AND (:type IS NULL OR p.type = :type) " +
             "AND ((:searchBy = 'title' AND LOWER(p.title) LIKE LOWER(CONCAT('%', :search, '%'))) " +
             "OR (:searchBy = 'content' AND LOWER(p.content) LIKE LOWER(CONCAT('%', :search, '%'))) " +
